Added support for OS X
- Fixed OS X warning on size_t to uint32_t conversion
- Modified podspec file to support same platforms as AFNetworking

+++ AFDownloadRequestOperation.m
@@ -319,7 +319,7 @@ typedef void (^AFURLConnectionProgressiveOperationProgressBlock)(AFDownloadReque
 + (NSString *)md5StringForString:(NSString *)string {
     const char *str = [string UTF8String];
     unsigned char r[CC_MD5_DIGEST_LENGTH];
-    CC_MD5(str, strlen(str), r);
+    CC_MD5(str, (uint32_t)strlen(str), r);
     return [NSString stringWithFormat:@"%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x",
             r[0], r[1], r[2], r[3], r[4], r[5], r[6], r[7], r[8], r[9], r[10], r[11], r[12], r[13], r[14], r[15]];
 }
